v3.5.0
1、优化ShareSDKConnector模块架构，根据平台进行分包。
2、新增支付宝朋友圈分享平台。
3、ShareSDK模块接口迁移：原ShareSDK中的获取好用，添加好友，调用网络API三种功能接口从ShareSDK.framewrok迁移
到ShareSDKExtention.framework
4、ShareSDK模块中的设置邮件、SMS和Google+的风格的接口移到平台的Connector模块。
5、使用配置文件分享优化成独立模块ShareSDKConfigFile。
6、更新了新浪微博SDK。
7、修复默认菜单在iOS10上部分平台(如：微信朋友圈)中名称显示不全的问题。
8、Facebook：修复分享WebPage类型是URL和网络图片URL可能被截断的问题；修复Facebook API
v2.8的应用在获取用户信息时的报错问题
9、修复Twitter在获取好友时没有反应的问题。

以下用户建议更新：
1、对包体大小要求较高的用户。
2、需要支持支付宝朋友圈的用户。
3、需要解决默认分享菜单栏平台名字显示不全问题的用户。
4、集成了Facebook且使用了WebPage分享类型的用户。
5、集成了Facebook且使用了v2.8版本Facebook开放平台应用的用户。
6、集成Twitter平台且需要获取好友信息的用户

+++ ShareSDK/Support/Required/MOBFoundation.framework/Headers/MOBFHttpService.h
@@ -51,7 +51,7 @@ typedef void(^MOBFHttpFaultEvent) (NSError *error);
  *  @param totalBytes  总字节数
  *  @param loadedBytes 上传字节数据
  */
-typedef void(^MOBFHttpUploadProgressEvent) (NSInteger totalBytes, NSInteger loadedBytes);
+typedef void(^MOBFHttpUploadProgressEvent) (int64_t totalBytes, int64_t loadedBytes);
 
 /**
  *  HTTP服务类
